Moved Book to a high order component folder
