+++ React/wbooks/src/app/components/Book/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import defaultBookIcon from '../../../../assets/default_book.svg';
-import { bookPropType } from '../../../../../redux/books/proptypes';
+import defaultBookIcon from '../../assets/default_book.svg';
+import { bookPropType } from '../../../redux/books/proptypes';
 
 import styles from './styles.scss';
