+++ React/wbooks/src/app/components/Book/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../../constants/colors.scss';
+@import '../../../constants/colors.scss';
 
 .bookContainer {
   display: flex;
