+++ React/wbooks/src/app/screens/Home/layout.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 
 import Loader from '../../components/Loader';
 import { bookArrayPropType } from '../../../redux/books/proptypes';
+import Book from '../../components/Book';
 
 import Combobox from './components/Combobox';
-import Book from './components/Book';
 import SearchInput from './components/SearchInput';
 import styles from './styles.scss';
